requirements.txt
playwright>=1.44.0
google-genai>=1.0.0
requests>=2.32.0
python-dotenv>=1.0.0
Pillow>=10.0.0
